requirements.txt
asgiref==3.5.1
Django==4.0.4
django-ckeditor==6.4.0
django-filter==21.1
django-js-asset==2.0.0
sqlparse==0.4.2
